Create README.md
using this site is like reading Latin.

I have  a problem with GRCT where the town  search is permanently lagging with the eternal spinning circle. It appears to be a known issue but I don't speak Latin or Geek. Can someone suggest a solution? Hate to give up the game over such a simple but impossibly irritating bug.